Unexport partial fields
- Move unit test to _internal
- Remove example_test
